feat(renderer): give the bed channel gain 0.1 dB resolution
The bed editor's per-channel gain moved in whole decibels while the
per-speaker output gain next to it moves in 0.1 dB — an inconsistency
with no reason behind it once traced: the decoder side is i8 by spec,
and that integer quietly propagated through the layout entry (i32), the
audio events (Option<i8>), the cached channel state and the display
objects, none of which are decoder-owned.

Carry the gain as f32 everywhere the renderer owns it: the layout
entry, the plan events, the planner's trim cache, the per-channel state
and the display objects. The decoder side keeps its i8 — bridge channel
gains and DAMF event gains convert at the boundary, so stream metadata
still arrives in whole dB and the fractional resolution belongs to the
bed trim summed on top. −128 stays the −inf convention, now a floor
(GAIN_DB_NEG_INF) with sums clamping onto it exactly, so the silence
test remains a single comparison; the shared gain_db_to_linear helper
replaces the three hand-rolled conversions that each knew about the
sentinel.

The layout file writes gain_db with the same 0.01 dB tolerance as the
render-config gains, so existing configs round-trip unchanged and a
default entry stays absent. The OSC object frame keeps sending the
display gain as a rounded Int: the native mpv overlay parses that tag,
and the editor reads the bed itself, not the display frame.

Studio's slider steps by 0.1 with one-decimal display, and the payload
carries the fractional value the engine now preserves. The event-path
regression tests exercise a fractional trim (−6.5 dB) end to end.

+++ omniphony-renderer/orender_engine/src/channel_objects.rs
@@ -167,7 +167,7 @@ impl ChannelObjectStages {
             .map(move |(index, spec)| SpatialChannelEvent {
                 channel_idx: channel_count + index,
                 is_bed: false,
-                gain_db: Some(spec.gain_db),
+                gain_db: Some(f32::from(spec.gain_db)),
                 ramp_length: Some(0),
                 size: Some(spec.size),
                 position: Some(spec.position),
@@ -190,7 +190,7 @@ impl ChannelObjectStages {
             z: spec.position[2] as f32,
             coord_mode: "cartesian".to_string(),
             direct_speaker_index: None,
-            gain: spec.gain_db as i32,
+            gain: f32::from(spec.gain_db),
             priority: 0.0,
             size: spec.size,
             fixed: false,
